extract GameLogic from layout plugin
This gives a cleaner interface between the UI / layout and the game
logic, which should make implementing the networking part much simpler.
At the same time, the LayoutState was small enough before and even shrank
further through this, and it is so tightly coupled with the layout
plugin that it did not seem useful to have it in a separate module.