use a tokio::sync::mpsc channel for playing cards
This does not fully work yet, but should implement the opposite
communication direction.  The result feels Frankenstein-like, because
we now use a std::sync::mpsc channel for the tokio WS->bevy direction,
and an async channel for the opposite direction.

Also, we broke the card covering, and all UIs still show the same state
(all hands open, and in same position).